requirements.txt
requests>=2.25.0
beautifulsoup4>=4.9.0
PyYAML>=5.4.0
Pillow>=8.0.0
